Remove DefaultStereotypeProvider
