+++ src/OrchardCore.Modules/OrchardCore.ContentTypes/Startup.cs
@@ -20,7 +20,6 @@ namespace OrchardCore.ContentTypes
             services.AddScoped<IPermissionProvider, Permissions>();
             services.AddScoped<INavigationProvider, AdminMenu>();
             services.AddScoped<IContentDefinitionService, ContentDefinitionService>();
-            services.AddScoped<IStereotypesProvider, DefaultStereotypesProvider>();
             services.AddScoped<IStereotypeService, StereotypeService>();
             services.AddScoped<IContentDefinitionDisplayHandler, ContentDefinitionDisplayCoordinator>();
             services.AddScoped<IContentDefinitionDisplayManager, DefaultContentDefinitionDisplayManager>();
